Fix empty contract data (revert utils move)
